use the Base version of func_for_method_checked
This package was noted to fail on the 1.2 branch of Julia. The usage of `func_for_method_checked` was changed in https://github.com/JuliaLang/julia/pull/31025 to also take an `sparams` argument. The old usage was deprecated but the function `depwarn` is not available in `Core.Compiler`. Using the `Base` version of this function will show the deprecation message instead of hard errorring.