fix: add support for comprehensions with conditionals
Comprehension support works by parsing the underlying generators, and
generators with `if` conditions have `filter` as their `Expr` form's
second argument. This previously wasn't supported, as the generator
parsing code expected the `for` iteration's `=` to be the second arg
instead. 

This change adds support for the case of generators with filters, and
adds relevant tests.